refactor: remove SW stuff; refactor lib/errors
