+++ index.ts
@@ -18,7 +18,7 @@ import {
   EncryptionType,
   SigningType,
 } from './src/lib/interactionManager/types'
-import { BackendError } from './src/lib/errors/types'
+import { SDKError, ErrorCode } from './src/lib/errors'
 
 export {
   ICredentialRequest as CredentialRequirements,
@@ -106,7 +106,7 @@ export class JolocomSDK extends BackendMiddleware {
     }
 
     if (!pass) {
-      if (!auto) throw new BackendError(BackendError.codes.NoWallet)
+      if (!auto) throw new SDKError(ErrorCode.NoWallet)
 
       console.warn('Generating a random password')
       pass = (await generateSecureRandomBytes(32)).toString('base64')
@@ -117,8 +117,8 @@ export class JolocomSDK extends BackendMiddleware {
       return await this.loadIdentity(storedDid)
     } catch (err) {
       if (
-        (!(err instanceof BackendError) ||
-          err.message !== BackendError.codes.NoWallet) &&
+        (!(err instanceof SDKError) ||
+          err.message !== ErrorCode.NoWallet) &&
         !auto
       ) {
         throw err
