+++ src/lib/errors.ts
@@ -1,25 +1,45 @@
-import { routeList } from '../routeList'
-import strings from '../locales/strings'
-import { ErrorCode, initErrorReporting, reportError } from './errors/index'
-export { ErrorCode, initErrorReporting, reportError }
+export enum ErrorCode {
+  Unknown = 'Unknown',
 
-export class AppError extends Error {
-  // private code: ErrorCode
-  public origError: any
-  public navigateTo: routeList
-  public code: ErrorCode
+  WalletInitFailed = 'WalletInit',
+  SaveClaimFailed = 'SaveClaim',
+  SaveExternalCredentialFailed = 'SaveExtCred',
+  SaveCredentialMetadataFailed = 'SaveCredMetadata',
+  TransportNotSupported = 'UnsuportedTransport',
 
-  public constructor(
-    code = ErrorCode.Unknown,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    origError?: any,
-    navigateTo: routeList = routeList.AppInit,
-  ) {
-    super(strings[code] || strings[ErrorCode.Unknown])
-    this.code = code
-    this.origError = origError
-    this.navigateTo = navigateTo
-  }
+  DeepLinkUrlNotFound = 'DeepLinkUrlNotFound',
+  TokenExpired = 'TokenExpired',
+  InvalidSignature = 'InvalidSignature',
+  WrongDID = 'WrongDID',
+  WrongNonce = 'WrongNonce',
+  WrongFlow = 'WrongFlow',
+
+  AuthenticationRequestFailed = 'AuthRequest',
+  AuthenticationResponseFailed = 'AuthResponse',
+
+  CredentialOfferFailed = 'CredOffer',
+  CredentialsReceiveFailed = 'CredsReceive',
+  CredentialRequestFailed = 'CredRequest',
+  CredentialResponseFailed = 'CredResponse',
+  ParseJWTFailed = 'ParseJWT',
+
+  RegistrationFailed = 'Registration',
+
+  InvalidToken = 'InvalidToken',
+
+  NoEntropy = 'NoEntropy',
+  NoKeyProvider = 'NoKeyProvider',
+  NoWallet = 'NoWallet',
+  DecryptionFailed = 'DecryptionFailed',
 }
 
-export const errorTitleMessages = [strings.DAMN, strings.OH_NO, strings.UH_OH]
+
+export class SDKError extends Error {
+  public static codes = ErrorCode
+  public origError?: Error
+
+  public constructor(code: ErrorCode, origError?: Error) {
+    super(code)
+    this.origError = origError
+  }
+}
