--- src/locales/de.json
@@ -1,119 +1,0 @@
-{
-    "Account is at risk": "Identität gefährdet",
-    "Add claim": "Weiter",
-    "Address Line1": "Anschrift",
-    "Address Line2": "Anschrift (optional)",
-    "All claims": "Alle Claims",
-    "Authentication request failed": "Authentifizierungsanfrage fehlgeschlagen",
-    "Authentication response failed": "Authentifizierungsantwort fehlgeschlagen",
-    "Authorization request": "Genehmigungsanfrage",
-    "Authorize": "Zulassen",
-    "Back to signup": "Zurück zur Registrierung",
-    "Backup your Identity": "Sichere deine Identität",
-    "Cancel": "Abbrechen",
-    "Choose the right word or press enter": "Wähle das richtige Wort und drücke Enter",
-    "City": "Stadt",
-    "Coming Soon": "Demnächst",
-    "Confirm": "Bestätigen",
-    "Confirm and check": "Bestätigen und überprüfen",
-    "Confirm payment": "Bezahlung Bestätigen",
-    "Contact": "Kontakt\n",
-    "Continue": "Weiter",
-    "Could not find receiving application": "Die empfangende App konnte nicht gefunden werden",
-    "Could not parse JSONWebToken": "JSONWebToken konnte nicht verarbeitet werden",
-    "Could not receive credentials": "Claim konnte nicht empfangen werden",
-    "Could not save claim": "Claim konnte nicht gespeichert werden",
-    "Could not save external credential": "Externer Claim konnte nicht gespeichert werden",
-    "Country": "Land\n",
-    "Credential offer failed": "Claimangebot fehlgeschlagen",
-    "Credential request failed": "Claimanfrage fehlgeschlagen",
-    "Credential response failed": "Claimantwort fehlgeschlagen",
-    "Damn": "Mist",
-    "Delete Identity": "Identität löschen",
-    "Deny": "Nein Danke",
-    "Document details/claims": "Dokumentdetails / Claims",
-    "Documents": "Dokumente",
-    "Done": "Fertig",
-    "Email": "Email",
-    "Encrypting and storing data locally": "Daten lokal verschlüsseln und speichern",
-    "Enhanced privacy": "Erhöhte Privatssphäre",
-    "Family Name": "Nachname",
-    "For": "Für",
-    "For security purposes, we need some randomness": "Aus Sicherheitsgründen brauchen wir zufällig erzeugte Daten",
-    "Forget about long forms and registrations": "Vergiss lange Formulare und Anmeldungen",
-    "Fueling with ether": "Mit Ether aufladen",
-    "Full phrase verification": "Vollständige Prüfung",
-    "Get started": "Loslegen",
-    "Give us a few moments": "Nur einen Moment bitte",
-    "Given Name": "Vorname",
-    "Go back": "Zurück",
-    "Greater control": "Mehr Kontrolle",
-    "If you have noted down your phrase, put the six given words on their right places": "Wenn du den Satz aufgeschrieben hast, platziere die Wörter an den richtigen Stellen",
-    "Initialization failed": "Initialisierung fehlgeschlagen",
-    "Instantly access services without using your social media profiles": "Sofortiger Zugang zu Diensten ohne Verwendung deiner Social-Media-Profile",
-    "It's easy": "Es ist einfach",
-    "Keep all your data with you in one place, available at any time": "Bewahre all deine Daten an einem Ort auf, jederzeit verfügbar",
-    "Language": "Sprache",
-    "Login records": "Login Einträge",
-    "Make it secure": "Sichere es",
-    "Mobile Phone": "Mobilnummer",
-    "My Identity": "Meine Identität",
-    "Name": "Name",
-    "Name of issuer": "Name des Ausstellers",
-    "No documents to see here": "Du hast noch keine Dokumente",
-    "No local claims": "Keine lokalen Claims",
-    "Oh no": "Oh Nein",
-    "Other": "Andere",
-    "Payment request failed": "Zahlungsanfrage fehlgeschlagen",
-    "Payment response failed": "Zahlungsantwort fehlgeschlagen",
-    "Personal": "Persönlich",
-    "Please tap the screen and draw on it randomly": "Tippe auf den Bildschirm und zeichne drauf los",
-    "Postal Address": "Adresse",
-    "Postal Code": "Postleitzahl",
-    "Preparing launch": "Start vorbereiten",
-    "Protect your digital self against fraud": "Schütze dein digitales Ich gegen Betrug",
-    "Receiving new credential": "Neue Claims erhalten",
-    "Recover identity": "Identität wiederherstellen",
-    "Recovery": "Wiederherstellung",
-    "Registering decentralized identity": "Dezentrale Identität registrieren",
-    "Registration failed": "Registrierung fehlgeschlagen",
-    "Restore account": "Konto wiederherstellen",
-    "Self-signed": "Selbst-signiert",
-    "Set up a secure phrase to recover your account in the future if your phone is stolen or is damaged": "Erstelle einen sicheren Satz um deine Identität wiederherzustellen falls dein Smartphone mal gestohlen wird oder kaputt geht",
-    "Settings": "Einstellungen",
-    "Share claims": "Claims teilen",
-    "Share only the information a service really needs": "Teile nur die Informationen, die ein Dienst wirklich benötigt",
-    "Show my phrase again": "Zeig mir den Satz noch mal",
-    "Start writing your seed-phrase and it will appears here word by word": "Beginne deinen Satz zur Wiederherstellung zu schreiben. Wort für Wort",
-    "Take back control of your digital self and protect your private data against unfair usage": "Übernimm die Kontrolle über dein digitales Ich und schütze deine privaten Daten gegen schlechte Nutzung",
-    "Tap here": "Hier tippen",
-    "Telephone": "Telefon",
-    "The order was not correct": "Die Reihenfolge war nicht korrekt",
-    "The word is not correct, check for typos": "Das Word ist nicht korrekt.",
-    "There was an error with your request": "Bei deiner Anfrage ist ein Fehler aufgetreten",
-    "This service is asking you to share the following claims": "Dieser Dienst bittet dich, die folgenden Claims zu teilen",
-    "To": "An",
-    "Track where you sign in to services": "Verfolge dein Anmeldeverhalten",
-    "Try again with another six words from your secure phrase": "Versuch es noch mal mit anderen Wörtern aus deinem Satz",
-    "Uh oh": "Hmmm",
-    "Unable to initialize wallet": "Wallet konnte nicht initialisiert werden",
-    "Unknown Error": "Unbekannter Fehler",
-    "Without these words, you cannot access your wallet again": "Ohne diese Wörter kannst du nicht mehr auf deine SmartWallet zugreifen",
-    "Would you like to": "Würdest du gern",
-    "Write these words down on an analog and secure place": "Schreibe diese Wörter an einem analogen und sicheren Ort auf",
-    "Yes, I wrote it down": "Ja, ich habe es aufgeschrieben",
-    "You can scan the qr code now!": "Scan die QR-Code",
-    "You haven't logged in to any services yet": "Du hast dich bisher bei keinem Service eingeloggt",
-    "Your Jolocom Wallet": "Deine Jolocom Wallet",
-    "Your identity is already backed up": "Deine Identität ist bereits gesichert",
-    "Your preferences": "Deine Einstellungen",
-    "add": "hinzufügen",
-    "completed": "vollständig",
-    "create\nnew account": "neue Identität erstellen ",
-    "expired": "Abgelaufen",
-    "if it is your first experience\nof Jolocom SmartWallet": "Wenn das deine erste Erfahrung\nmit der Jolocom SmartWallet ist",
-    "recover your\ndigital identity": "stelle deine digitale Identität wieder her",
-    "to set up your identity": "Deine Identität wird erstellt",
-    "version": "Version",
-    "with your SmartWallet?": "mit deiner SmartWallet?"
-}
