--- src/locales/nl.json
@@ -1,119 +1,0 @@
-{
-    "Account is at risk": "Account is at risk",
-    "Add claim": "Claim toevoegen",
-    "Address Line1": "Adres regel 1",
-    "Address Line2": "Adres regel 2",
-    "All claims": "Alle claims",
-    "Authentication request failed": "Authentication request failed",
-    "Authentication response failed": "Authentication response failed",
-    "Authorization request": "Authorization request",
-    "Authorize": "Authorize",
-    "Back to signup": "Back to signup",
-    "Backup your Identity": "Backup your Identity",
-    "Cancel": "Cancel",
-    "Choose the right word or press enter": "Choose the right word or press enter",
-    "City": "Stad",
-    "Coming Soon": "Binnenkort",
-    "Confirm": "Bevestigen",
-    "Confirm and check": "Confirm and check",
-    "Confirm payment": "Bevestig betaling",
-    "Contact": "Contact",
-    "Continue": "Doorgaan",
-    "Could not find receiving application": "Could not find receiving application",
-    "Could not parse JSONWebToken": "Could not parse JSONWebToken",
-    "Could not receive credentials": "Could not receive credentials",
-    "Could not save claim": "Could not save claim",
-    "Could not save external credential": "Could not save external credential",
-    "Country": "Land",
-    "Credential offer failed": "Credential offer failed",
-    "Credential request failed": "Credential request failed",
-    "Credential response failed": "Credential response failed",
-    "Damn": "Damn",
-    "Delete Identity": "Delete Identity",
-    "Deny": "Weigeren",
-    "Document details/claims": "Documentdetails/claims",
-    "Documents": "Documenten",
-    "Done": "Done",
-    "Email": "E-mail",
-    "Encrypting and storing data locally": "Data lokaal versleutelen en opslaan",
-    "Enhanced privacy": "Verhoogde privacy",
-    "Family Name": "Achternaam",
-    "For": "Voor",
-    "For security purposes, we need some randomness": "Voor de veiligheid hebben we wat willekeur nodig",
-    "Forget about long forms and registrations": "Zonder lange formulieren en registraties",
-    "Fueling with ether": "Ether tanken",
-    "Full phrase verification": "Full phrase verification",
-    "Get started": "Begin",
-    "Give us a few moments": "Geef ons een moment",
-    "Given Name": "Gegeven naam",
-    "Go back": "Terug",
-    "Greater control": "Meer controle",
-    "If you have noted down your phrase, put the six given words on their right places": "If you have noted down your phrase, put the six given words on their right places",
-    "Initialization failed": "Initialization failed",
-    "Instantly access services without using your social media profiles": "Direct toegang tot services, niet via uw sociale media-profielen",
-    "It's easy": "Simpel",
-    "Keep all your data with you in one place, available at any time": "Bewaar al uw gegevens op één plaats, altijd beschikbaar",
-    "Language": "Taal",
-    "Login records": "Login records",
-    "Make it secure": "Make it secure",
-    "Mobile Phone": "Mobiel nummer",
-    "My Identity": "Mijn identiteit",
-    "Name": "Naam",
-    "Name of issuer": "Naam van uitgevende instantie",
-    "No documents to see here": "No documents to see here",
-    "No local claims": "Geen lokale claims",
-    "Oh no": "Oh no",
-    "Other": "Other",
-    "Payment request failed": "Payment request failed",
-    "Payment response failed": "Payment response failed",
-    "Personal": "Persoonlijk",
-    "Please tap the screen and draw on it randomly": "Tik op het scherm en teken er willekeurig op",
-    "Postal Address": "Postadres",
-    "Postal Code": "Postcode",
-    "Preparing launch": "De lancering voorbereiden",
-    "Protect your digital self against fraud": "Bescherm uw digitale identiteit tegen fraude",
-    "Receiving new credential": "Nieuwe inloggegevens ontvangen",
-    "Recover identity": "Recover identity",
-    "Recovery": "Recovery",
-    "Registering decentralized identity": "Gedecentraliseerde identiteit registreren",
-    "Registration failed": "Registration failed",
-    "Restore account": "Restore account",
-    "Self-signed": "Zelfondertekend",
-    "Set up a secure phrase to recover your account in the future if your phone is stolen or is damaged": "Set up a secure phrase to recover your account in the future if your phone is stolen or is damaged",
-    "Settings": "Instellingen",
-    "Share claims": "Deel claims",
-    "Share only the information a service really needs": "Deel alleen de informatie die een service echt nodig heeft",
-    "Show my phrase again": "Show my phrase again",
-    "Start writing your seed-phrase and it will appears here word by word": "Start writing your seed-phrase and it will appears here word by word",
-    "Take back control of your digital self and protect your private data against unfair usage": "Neem de controle over uw digitale zelf over en bescherm uw privégegevens tegen oneerlijk gebruik",
-    "Tap here": "Tap here",
-    "Telephone": "Telefoon",
-    "The order was not correct": "The order was not correct",
-    "The word is not correct, check for typos": "The word is not correct, check for typos",
-    "There was an error with your request": "Er is een fout opgetreden bij uw aanvraag",
-    "This service is asking you to share the following claims": "Deze service vraagt ​​u om de volgende claims te delen",
-    "To": "Naar",
-    "Track where you sign in to services": "Volg bij welke services u zich aanmeldt",
-    "Try again with another six words from your secure phrase": "Try again with another six words from your secure phrase",
-    "Uh oh": "Uh oh",
-    "Unable to initialize wallet": "Unable to initialize wallet",
-    "Unknown Error": "Unknown Error",
-    "Without these words, you cannot access your wallet again": "Zonder deze woorden heeft u geen toegang meer tot uw portefeuille",
-    "Would you like to": "Would you like to",
-    "Write these words down on an analog and secure place": "Schrijf deze woorden op een analoge en veilige plaats op",
-    "Yes, I wrote it down": "Ja, ik heb het opgeschreven",
-    "You can scan the qr code now!": "U kunt de QR-code nu scannen!",
-    "You haven't logged in to any services yet": "You haven't logged in to any services yet",
-    "Your Jolocom Wallet": "Uw Jolocom Wallet",
-    "Your identity is already backed up": "Your identity is already backed up",
-    "Your preferences": "Jouw Voorkeuren",
-    "add": "voeg toe",
-    "completed": "completed",
-    "create\nnew account": "create\nnew account",
-    "expired": "expired",
-    "if it is your first experience\nof Jolocom SmartWallet": "if it is your first experience\nof Jolocom SmartWallet",
-    "recover your\ndigital identity": "recover your\ndigital identity",
-    "to set up your identity": "om je identiteit in te stellen",
-    "version": "versie",
-    "with your SmartWallet?": "with your SmartWallet?"
-}
