--- src/polyfills/i18n.ts
@@ -1,4 +1,0 @@
-export default {
-  // src/locales/i18n.ts default export
-  t: (str: string) => str,
-}
