--- src/polyfills/reactNative.ts
@@ -1,3 +1,0 @@
-import { randomBytes } from 'crypto'
-
-export const NativeModules = { RNRandomBytes: { randomBytes } }
