--- src/polyfills/rnSplashScreen.ts
@@ -1,8 +1,0 @@
-function logCall(pref: string) {
-  return (msg: string) => console.log(`${pref}: ${msg}`)
-}
-
-export default {
-  // react-native-splash-screen
-  hide: () => logCall('SplashScreen')('hide'),
-}
